Reuse encoder_video from LTX2 for Cosmos3
