+++ examples/cosmos3/inference_cosmos3.py
@@ -29,8 +29,7 @@ import torch
 from huggingface_hub import snapshot_download
 
 from diffusers import Cosmos3OmniDiffusersPipeline
-from diffusers.pipelines.cosmos.export_utils import save_wav
-from diffusers.utils import export_to_video, load_image
+from diffusers.utils import encode_video, export_to_video, load_image
 
 
 HF_REPO = "nvidia/Cosmos3-Nano"
@@ -92,15 +91,19 @@ def main():
         result.video[0].save(save_path, format="JPEG", quality=85)
     else:
         save_path = output_dir / "sample.mp4"
-        # macro_block_size=1 allows arbitrary frame sizes (Cosmos3 outputs are not always divisible by 16).
-        export_to_video(result.video, str(save_path), fps=int(args.fps), quality=10, macro_block_size=1)
+        if result.sound is not None:
+            assert pipeline.sound_tokenizer is not None
+            encode_video(
+                result.video,
+                fps=int(args.fps),
+                audio=result.sound,
+                audio_sample_rate=pipeline.sound_tokenizer.sample_rate,
+                output_path=str(save_path),
+            )
+        else:
+            # macro_block_size=1 allows arbitrary frame sizes (Cosmos3 outputs are not always divisible by 16).
+            export_to_video(result.video, str(save_path), fps=int(args.fps), quality=10, macro_block_size=1)
     print(f"Saved: {save_path}")
-
-    if result.sound is not None:
-        assert pipeline.sound_tokenizer is not None
-        wav_path = output_dir / "sample.wav"
-        save_wav(result.sound, wav_path, pipeline.sound_tokenizer.sample_rate)
-        print(f"Saved: {wav_path}")
 
 
 if __name__ == "__main__":
